Created script to download problems
Created script which downloads the problem pages from Project Euler website
and creates locally copies of them in problems folder.
The local copies are markdown files with the problem details in and links
to the solution script.
The solution scripts have all been renamed to follow the naming convention
of the markdown problem files.
Updated gitignore to not include anything in the problems folder as these
are now downloaded via the script.
Added requests and beautifulsoup4 packages to the environment.